+++ r-aks-monitoring.tf
@@ -1,0 +1,49 @@
+resource "azurerm_monitor_data_collection_rule" "dcr" {
+  for_each            = var.data_collection_rule_enabled ? toset(["enabled"]) : []
+  name                = local.dcr_name
+  location            = var.location
+  resource_group_name = var.resource_group_name
+
+  destinations {
+    log_analytics {
+      name                  = "default-workspace"
+      workspace_resource_id = var.data_collection_rule.log_analytics_workspace_id
+    }
+  }
+
+  data_flow {
+    destinations = ["default-workspace"]
+    streams      = var.data_collection_rule.data_streams
+  }
+
+  data_sources {
+    extension {
+      name           = "ContainerInsightsExtension"
+      extension_name = "ContainerInsights"
+      streams        = var.data_collection_rule.data_streams
+      extension_json = jsonencode({
+        dataCollectionSettings = {
+          enableContainerLogV2   = var.data_collection_rule.container_log_v2_enabled
+          interval               = var.data_collection_rule.data_collection_interval
+          namespaceFilteringMode = var.data_collection_rule.namespace_filtering_mode
+          namespaces             = var.data_collection_rule.namespaces_filter
+        }
+      })
+    }
+  }
+  lifecycle {
+    precondition {
+      condition     = var.data_collection_rule.log_analytics_workspace_id != null
+      error_message = "You must provide log_analytics_workspace_id in data_collection_rule variable."
+    }
+  }
+
+  tags = merge(local.default_tags, var.extra_tags)
+}
+
+resource "azurerm_monitor_data_collection_rule_association" "dcr" {
+  for_each                = var.data_collection_rule_enabled ? toset(["enabled"]) : []
+  name                    = azurerm_kubernetes_cluster.aks.name
+  target_resource_id      = azurerm_kubernetes_cluster.aks.id
+  data_collection_rule_id = azurerm_monitor_data_collection_rule.dcr["enabled"].id
+}